bitget_bot/requirements.txt
pandas
ccxt
constantly
numpy
ta
datetime
configparser
mysql-connector-python
telegram_send
#tuto telegram bot https://medium.com/@robertbracco1/how-to-write-a-telegram-bot-to-send-messages-with-python-bcdf45d0a580


